moved AddAvatar from pages folder to components folder
